% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scrape_hsd.r
\name{scrape_hsd}
\alias{scrape_hsd}
\title{Title}
\usage{
scrape_hsd(state_name, state_abbreviation)
}
\arguments{
\item{state_name}{the state's name}

\item{state_abbreviation}{the state's two-character abbreviation}
}
\value{
a data frame
}
\description{
Title
}
\examples{
wyoming<-scrape_hsd('Wyoming','WY')
}
